Small fixes
- Move Redis keys per server to avoid cleaning up wrong peers, leaving
stale entries.
- Shuffle around for readability, cleanup.
